# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ClipAPI
  version: 0.1.0
  description: >
    REST contract for timed short-video transcripts. Errors never invent
    caption lines and never charge credits.
servers:
  - url: https://api.clipapi.dev
    description: Production placeholder
paths:
...
  /v1/billing/checkout:
    post:
      operationId: createCheckout
      summary: Start Stripe Checkout for the monthly $5 plan
      description: >
        0 credits. Returns a Checkout URL for the monthly plan (1,000 credits
        at $5). Without a configured Stripe client the server fails closed and
        does not call Stripe. Credits are granted only by the signed webhook.
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/billing/webhook:
    post:
      operationId: stripeWebhook
      summary: Apply a signed Stripe event
      security: []
      description: >
        Stripe-signed JSON. checkout.session.completed at $5 (500 cents)
        adds 1,000 credits and leaves unused free credits in place.
        invoice.payment_failed locks the key to free remaining 0.
        Invalid signatures return 400. Not bearer-auth.
      responses:
        "200":
          description: Event applied or replayed
        "400":
          description: Invalid signature or event
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: ck_live_... or ck_test_...
  schemas:
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - unauthorized
        - payment_required
        - not_found
        - no_transcript
        - unsupported_platform
        - rate_limited
        - upstream_blocked
        - internal
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        retryable:
          type: boolean
    ErrorMeta:
      type: object
      additionalProperties: false
      required: [creditsCharged, requestId]
      properties:
        creditsCharged:
          type: integer
          const: 0
        requestId:
          type: string
          pattern: "^req_"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          $ref: "#/components/schemas/ErrorMeta"
    SuccessMeta:
      type: object
      additionalProperties: false
      required: [cached, creditsCharged, requestId, upstreamMs]
      properties:
        cached:
          type: boolean
        creditsCharged:
          type: integer
          minimum: 0
        requestId:
          type: string
          pattern: "^req_"
        upstreamMs:
          type: integer
          minimum: 0
...
    CheckoutData:
      type: object
      additionalProperties: false
      required: [url, sessionId, plan, amountCents, currency, credits]
      properties:
        url:
          type: string
        sessionId:
          type: string
        plan:
          type: string
          const: monthly
        amountCents:
          type: integer
          const: 500
        currency:
          type: string
          const: usd
        credits:
          type: integer
          const: 1000
    CheckoutEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/CheckoutData"
        meta:
          $ref: "#/components/schemas/SuccessMeta"
...
  responses:
...
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Internal:
      description: Server bug
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
security:
  - bearerAuth: []